Filament and AuthController

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -12,8 +13,13 @@ Route::get('/user', function (Request $request) {
 Route::prefix("/categories")->group(function () {
     Route::get("/", [CategoryController::class, "getAllCategories"]);
     Route::get("/{id}", [CategoryController::class, "getCategoryById"]);
-    Route::post("/", [CategoryController::class, "addCategory"]);
-    Route::delete("/{id}", [CategoryController::class, "deleteCategory"]);
-    Route::put("/edit/{id}", [CategoryController::class, "editCategory"]);
+    Route::post("/", [CategoryController::class, "addCategory"])->middleware("isItAdmin");
+    Route::delete("/{id}", [CategoryController::class, "deleteCategory"])->middleware("isItAdmin");
+    Route::put("/edit/{id}", [CategoryController::class, "editCategory"])->middleware("isItAdmin");
 });
 
+Route::prefix("/auth")->group(function(){
+    Route::post("/login", [AuthController::class, "login"]);
+    Route::post("/register", [AuthController::class, "register"]);
+    Route::post("/edituser", [AuthController::class, "editUser"])->middleware("verifyToken");
+});
